handle only postive

+++ string.c
@@ -1,6 +1,6 @@
 nclude <stdio.h>
 
-int my_strlen(char *s)
+size_t  my_strlen(char *s)
 {
 	   char *p = s;
